fix comunication module
