+++ .npmbundlerrc
@@ -1,4 +1,5 @@
 {
+	"sources": ["src", "assets"],
 	"create-jar": {
 		"output-dir": "dist",
 		"features": {
@@ -6,5 +7,12 @@
 			"web-context": "/banner"
 		}
 	},
+	"rules": [
+		{
+			"test": "\\.css$",
+			"exclude": "node_modules",
+			"use": ["style-loader"]
+		}
+	],
 	"dump-report": true
 }
