+++ .webpack/webpack.config.js
@@ -21,7 +21,7 @@ module.exports = {
                 },
             },
 			{
-				test: /\.css$/i,
+				test: /\.css$/,
 				use: [
 				  "style-loader",
 				  "css-loader",
