+++ assets/css/styles.css
@@ -10,4 +10,3 @@
 .pre {
 	font-family: monospace; white-space: pre;
 }
-
