+++ src/components/PortletComponent.css
@@ -1,0 +1,3 @@
+p {
+    background: red;
+}
