--- tailwind.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-    content: [
-      './src/**/*.{html,js,jsx,ts,tsx}', // Ajusta el path según tu estructura
-    ],
-    theme: {
-      extend: {},
-    },
-    plugins: [],
-  }
